backend/requirements.txt
asgiref
sklearn
scikit-learn
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-multiupload==0.6.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.11.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.2.0
pip==23.0.1
platformdirs==3.2.0
pycparser==2.21
PyJWT
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==65.5.1
simplejson==3.19.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.21.0
wheel==0.38.4